fix: No result glitch/infinite loop
