+++ src/components/recommend/SCSS/recommend.scss
@@ -31,7 +31,7 @@
   }
 
   .auc-Recommend-item {
-    width: 15%;
+    width: 20%;
   }
 
   .relatedItem {
