+++ src/components/searchbox/components/SearchInCategory.jsx
@@ -1,8 +1,7 @@
-// Component for the searchbox being used when a category is selected eg. 'Mens'
-import { queryAtom } from '@/config/searchboxConfig';
+// Component for the searchbox being used when a category is selected eg. 'Mens'xz
 import { useNavigate } from 'react-router-dom';
 import { CloseButton } from '@/assets/svg/SvgIndex';
-import { useRecoilValue } from 'recoil';
+
 const SearchInCategory = ({ state }) => {
   // navigate is used by React Router
   const navigate = useNavigate();
