+++ src/components/searchresultpage/SCSS/searchResultsPage.scss
@@ -84,7 +84,7 @@
 // No Results
 // ------------------------------------------
 .no-results {
-  width: 80%;
+  width: 95%;
   height: 80vh;
   margin: 0rem auto;
   padding: 2rem 2rem;
@@ -116,6 +116,7 @@
             display: flex;
             align-items: center;
             gap: 1rem;
+            cursor: pointer;
 
             svg {
               margin-top: 0.2rem;
